Use np.integer instead of np.int in bokeh.protocol
This is apparently needed for Python 3 compatibility. Otherwise
you will get errors about np.int64 not being serializable, etc.

+++ bokeh/protocol.py
@@ -62,7 +62,7 @@ class BokehJSONEncoder(json.JSONEncoder):
             return obj.value / millifactor  #nanosecond to millisecond
         elif isinstance(obj, np.float):
             return float(obj)
-        elif isinstance(obj, np.int):
+        elif isinstance(obj, np.integer):
             return int(obj)
         # Datetime, Date
         elif isinstance(obj, (dt.datetime, dt.date)):
